pkg/cdi: add functions for writing Spec files.
Add Spec.Write() and Cache.WriteSpec() for writing Spec
files (atomically from the Cache refresh perspective).
Spec.Write writes the Spec to the file associated with
the Spec during its instantiation (either with NewSpec
or with ReadSpec). Cache.WriteSpec writes the Spec file
into the last of the configured Spec directories of the
Cache itself.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>