pkg/cdi: add functions for transient Spec handling.
Add convenience functions for dealing with transient Spec
files. A transient Spec is one whose lifecycle is tied to
events external to a CDI client, for instance the creation
and removal of a given container.

Transient Specs should be generated and eventually removed
using CreateTransientSpec() and RemoveTransientSpec(). This
should provide a consistent naming convention and guarantee
uniqueness across the generated file names.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>